+++ core/src/main/java/io/grpc/internal/DelayedClientTransport.java
@@ -21,7 +21,6 @@ import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.SettableFuture;
 import io.grpc.CallOptions;
 import io.grpc.Context;
-import io.grpc.InternalLogId;
 import io.grpc.InternalTransportStats;
 import io.grpc.LoadBalancer.PickResult;
 import io.grpc.LoadBalancer.PickSubchannelArgs;
@@ -48,7 +47,7 @@ import javax.annotation.concurrent.GuardedBy;
  * thus the delayed transport stops owning the stream.
  */
 final class DelayedClientTransport implements ManagedClientTransport {
-  private final InternalLogId lodId = InternalLogId.allocate(getClass().getName());
+  private final LogId lodId = LogId.allocate(getClass().getName());
 
   private final Object lock = new Object();
 
@@ -338,7 +337,7 @@ final class DelayedClientTransport implements ManagedClientTransport {
 
   // TODO(carl-mastrangelo): remove this once the Subchannel change is in.
   @Override
-  public InternalLogId getLogId() {
+  public LogId getLogId() {
     return lodId;
   }
