+++ core/src/main/java/io/grpc/internal/ForwardingConnectionClientTransport.java
@@ -19,7 +19,6 @@ package io.grpc.internal;
 import com.google.common.util.concurrent.ListenableFuture;
 import io.grpc.Attributes;
 import io.grpc.CallOptions;
-import io.grpc.InternalLogId;
 import io.grpc.InternalTransportStats;
 import io.grpc.Metadata;
 import io.grpc.MethodDescriptor;
@@ -54,7 +53,7 @@ abstract class ForwardingConnectionClientTransport implements ConnectionClientTr
   }
 
   @Override
-  public InternalLogId getLogId() {
+  public LogId getLogId() {
     return delegate().getLogId();
   }
