+++ core/src/main/java/io/grpc/internal/ManagedClientTransport.java
@@ -16,7 +16,6 @@
 
 package io.grpc.internal;
 
-import io.grpc.InternalWithLogId;
 import io.grpc.Status;
 import javax.annotation.CheckReturnValue;
 import javax.annotation.Nullable;
@@ -34,7 +33,7 @@ import javax.annotation.concurrent.ThreadSafe;
  * {@link Listener#transportTerminated}.
  */
 @ThreadSafe
-public interface ManagedClientTransport extends ClientTransport, InternalWithLogId {
+public interface ManagedClientTransport extends ClientTransport, WithLogId {
 
   /**
    * Starts transport. This method may only be called once.
