+++ core/src/main/java/io/grpc/internal/ServerTransport.java
@@ -16,13 +16,12 @@
 
 package io.grpc.internal;
 
-import io.grpc.InternalInstrumented;
 import io.grpc.InternalTransportStats;
 import io.grpc.Status;
 import java.util.concurrent.ScheduledExecutorService;
 
 /** An inbound connection. */
-public interface ServerTransport extends InternalInstrumented<InternalTransportStats> {
+public interface ServerTransport extends Instrumented<InternalTransportStats> {
   /**
    * Initiates an orderly shutdown of the transport. Existing streams continue, but new streams will
    * eventually begin failing. New streams "eventually" begin failing because shutdown may need to
