+++ core/src/main/java/io/grpc/internal/WithLogId.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016, gRPC Authors All rights reserved.
+ * Copyright 2017, gRPC Authors All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,15 +14,18 @@
  * limitations under the License.
  */
 
-package io.grpc;
+package io.grpc.internal;
 
 /**
- * Do not use this.
- *
- * <p>An object that has an ID that is unique within the JVM, primarily for debug logging.
+ * A loggable ID, unique for the duration of the program.
  */
-@Internal
-public interface InternalWithLogId extends WithLogId {
-  @Override
-  InternalLogId getLogId();
+public interface WithLogId {
+  /**
+   * Returns an ID that is primarily used in debug logs. It usually contains the class name and a
+   * numeric ID that is unique among the instances.
+   *
+   * <p>The subclasses of this interface usually want to include the log ID in their {@link
+   * #toString} results.
+   */
+  LogId getLogId();
 }
