+++ core/src/test/java/io/grpc/internal/ServerImplTest.java
@@ -51,7 +51,6 @@ import io.grpc.Context;
 import io.grpc.Grpc;
 import io.grpc.HandlerRegistry;
 import io.grpc.IntegerMarshaller;
-import io.grpc.InternalLogId;
 import io.grpc.InternalTransportStats;
 import io.grpc.Metadata;
 import io.grpc.MethodDescriptor;
@@ -1265,7 +1264,7 @@ public class ServerImplTest {
     }
 
     @Override
-    public InternalLogId getLogId() {
+    public LogId getLogId() {
       throw new UnsupportedOperationException();
     }
