+++ grpclb/src/main/java/io/grpc/grpclb/GrpclbLoadBalancer.java
@@ -22,13 +22,13 @@ import com.google.common.annotations.VisibleForTesting;
 import io.grpc.Attributes;
 import io.grpc.ConnectivityStateInfo;
 import io.grpc.EquivalentAddressGroup;
-import io.grpc.InternalLogId;
-import io.grpc.InternalWithLogId;
 import io.grpc.LoadBalancer;
 import io.grpc.Status;
 import io.grpc.grpclb.GrpclbConstants.LbPolicy;
 import io.grpc.internal.GrpcAttributes;
+import io.grpc.internal.LogId;
 import io.grpc.internal.ObjectPool;
+import io.grpc.internal.WithLogId;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -43,10 +43,10 @@ import javax.annotation.Nullable;
  * <p>Optionally, when requested by the naming system, will delegate the work to a local pick-first
  * or round-robin balancer.
  */
-class GrpclbLoadBalancer extends LoadBalancer implements InternalWithLogId {
+class GrpclbLoadBalancer extends LoadBalancer implements WithLogId {
   private static final Logger logger = Logger.getLogger(GrpclbLoadBalancer.class.getName());
 
-  private final InternalLogId logId = InternalLogId.allocate(getClass().getName());
+  private final LogId logId = LogId.allocate(getClass().getName());
 
   private final Helper helper;
   private final Factory pickFirstBalancerFactory;
@@ -82,7 +82,7 @@ class GrpclbLoadBalancer extends LoadBalancer implements InternalWithLogId {
   }
 
   @Override
-  public InternalLogId getLogId() {
+  public LogId getLogId() {
     return logId;
   }
