+++ grpclb/src/main/java/io/grpc/grpclb/GrpclbState.java
@@ -33,7 +33,6 @@ import io.grpc.Attributes;
 import io.grpc.ConnectivityState;
 import io.grpc.ConnectivityStateInfo;
 import io.grpc.EquivalentAddressGroup;
-import io.grpc.InternalLogId;
 import io.grpc.LoadBalancer.Helper;
 import io.grpc.LoadBalancer.PickResult;
 import io.grpc.LoadBalancer.PickSubchannelArgs;
@@ -43,6 +42,7 @@ import io.grpc.ManagedChannel;
 import io.grpc.Metadata;
 import io.grpc.Status;
 import io.grpc.grpclb.LoadBalanceResponse.LoadBalanceResponseTypeCase;
+import io.grpc.internal.LogId;
 import io.grpc.stub.StreamObserver;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
@@ -92,7 +92,7 @@ final class GrpclbState {
       }
     };
 
-  private final InternalLogId logId;
+  private final LogId logId;
   private final String serviceName;
   private final Helper helper;
   private final TimeProvider time;
@@ -130,7 +130,7 @@ final class GrpclbState {
       Helper helper,
       TimeProvider time,
       ScheduledExecutorService timerService,
-      InternalLogId logId) {
+      LogId logId) {
     this.helper = checkNotNull(helper, "helper");
     this.time = checkNotNull(time, "time provider");
     this.timerService = checkNotNull(timerService, "timerService");
